final project 11.12.21
* added create new loan view model

+++ app/src/main/java/com/w4eret1ckrtb1tch/homework/domain/entity/LoanRequest.kt
@@ -2,10 +2,11 @@ package com.w4eret1ckrtb1tch.homework.domain.entity
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
 data class LoanRequest(
-    val amount: Long,
+    val amount: BigDecimal,
     val firstName: String,
     val lastName: String,
     val percent: Double,
